Avances en foto de facebook al registrar

+++ src/Loogares/UsuarioBundle/Security/User/FacebookProvider.php
@@ -131,6 +131,7 @@ class FacebookProvider implements UserProviderInterface
                     if(file_put_contents('assets/images/temp/'.$fn, $result)) {                        
                         if(getimagesize('assets/images/temp/'.$fn)) {
                             $imagen = new UploadedFile('assets/images/temp/'.$fn, $fn);
+                            $user->setImagenFull("default.gif");
                             $user->file = $imagen;
                         }                        
                     }
@@ -141,9 +142,9 @@ class FacebookProvider implements UserProviderInterface
             
             $em->flush();
 
-            /*if(isset($fn)) {
+            if(isset($fn)) {
                 unlink('assets/images/temp/'.$fn);
-            }*/
+            }
         }
 
         if (empty($user)) {
